+++ src/main/java/de/marhali/easyi18n/io/parser/json5/Json5ArrayMapper.java
@@ -5,6 +5,7 @@ import de.marhali.easyi18n.util.StringUtil;
 import de.marhali.json5.Json5;
 import de.marhali.json5.Json5Array;
 import de.marhali.json5.Json5Primitive;
+
 import org.apache.commons.lang3.math.NumberUtils;
 
 import java.io.IOException;