+++ src/main/java/de/marhali/easyi18n/io/parser/json5/Json5Mapper.java
@@ -6,6 +6,7 @@ import de.marhali.easyi18n.util.StringUtil;
 import de.marhali.json5.Json5Element;
 import de.marhali.json5.Json5Object;
 import de.marhali.json5.Json5Primitive;
+
 import org.apache.commons.lang3.math.NumberUtils;
 import org.apache.commons.text.StringEscapeUtils;