+++ src/main/java/de/marhali/easyi18n/io/parser/json/JsonMapper.java
@@ -7,6 +7,7 @@ import com.google.gson.JsonPrimitive;
 import de.marhali.easyi18n.model.TranslationNode;
 import de.marhali.easyi18n.model.TranslationValue;
 import de.marhali.easyi18n.util.StringUtil;
+
 import org.apache.commons.lang3.math.NumberUtils;
 import org.apache.commons.text.StringEscapeUtils;