+++ src/main/java/de/marhali/easyi18n/io/parser/properties/PropertiesMapper.java
@@ -5,6 +5,7 @@ import de.marhali.easyi18n.model.KeyPath;
 import de.marhali.easyi18n.model.TranslationValue;
 import de.marhali.easyi18n.util.KeyPathConverter;
 import de.marhali.easyi18n.util.StringUtil;
+
 import org.apache.commons.lang3.math.NumberUtils;
 import org.apache.commons.text.StringEscapeUtils;